+++ deployments/kube/.gitignore
@@ -1,0 +1,2 @@
+auto/config/*
+yml/cega-users/*
