+++ deployments/kube/auto/configure.py
