+++ deployments/kube/auto/scripts/db.sql
