+++ deployments/kube/auto/scripts/defs.json
