+++ deployments/kube/auto/scripts/inbox.sh
