+++ deployments/kube/auto/scripts/mq.sh
