+++ deployments/kube/auto/scripts/rabbitmq.config
