--- deployments/kube/automate/.gitignore
@@ -1,1 +1,0 @@
-config/*
