--- deployments/kube/cega-mq/cm.cega-mq.yml
@@ -1,36 +1,0 @@
-apiVersion: v1
-kind: ConfigMap
-metadata:
-  name: cega-mq-config
-  namespace: localega
-data:
-  rabbitmq.config: |-
-    <rabbitmq.config>
-  defs.json: |-
-    <defs.json>
----
-apiVersion: v1
-kind: ConfigMap
-metadata:
-  name: cega-script
-  namespace: localega
-data:
-  cega-mq.sh: |-
-    #!/bin/bash
-
-    set -e
-
-    # Initialization
-    rabbitmq-plugins enable --offline rabbitmq_federation
-    rabbitmq-plugins enable --offline rabbitmq_federation_management
-    rabbitmq-plugins enable --offline rabbitmq_shovel
-    rabbitmq-plugins enable --offline rabbitmq_shovel_management
-
-    cp --remove-destination /temp/rabbitmq.config /etc/rabbitmq/rabbitmq.config
-    cp --remove-destination /temp/defs.json /etc/rabbitmq/defs.json
-    chown rabbitmq:rabbitmq /etc/rabbitmq/rabbitmq.config
-    chmod 640 /etc/rabbitmq/rabbitmq.config
-    chown rabbitmq:rabbitmq /etc/rabbitmq/defs.json
-    chmod 640 /etc/rabbitmq/defs.json
-
-    exec rabbitmq-server
