--- deployments/kube/cega-users/deploy.cega.yml
@@ -1,57 +1,0 @@
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: cega-users
-  namespace: localega
-  labels:
-    role: fake-users
-spec:
-  replicas: 1
-  selector:
-    matchLabels:
-      app: cega-users
-  template:
-    metadata:
-      labels:
-        app: cega-users
-        role: fake-users
-    spec:
-      containers:
-      - name: cega-users
-        image: python:3.6-alpine3.7
-        imagePullPolicy: IfNotPresent
-        command: ["/bin/sh","-c"]
-        args: ["pip install PyYAML aiohttp aiohttp_jinja2; python /script/server.py"]
-        env:
-          - name: USERS_PATH
-            value: /
-          - name: LEGA_INSTANCES
-            value: lega
-          - name: CEGA_REST_lega_PASSWORD
-            valueFrom:
-              secretKeyRef:
-                name: cega-user-secret
-                key: lega_password
-        ports:
-        - name: cega-users
-          containerPort: 80
-          protocol: TCP
-        volumeMounts:
-        - name: cega-users
-          mountPath: /cega/users
-        - name: cega-config
-          mountPath: /script
-      volumes:
-        - name: cega-config
-          configMap:
-            name: cega-users-config
-            items:
-              - key: users.html
-                path: users.html
-              - key: server.py
-                path: server.py
-            defaultMode: 0744
-        - name: cega-users
-          hostPath:
-            path: /
-            type: DirectoryOrCreate
