--- deployments/kube/cega-users/svc.cega.yml
@@ -1,15 +1,0 @@
-apiVersion: v1
-kind: Service
-metadata:
-  name: cega-users
-  namespace: localega
-  labels:
-    app: cega-users
-spec:
-  type: NodePort
-  ports:
-  - port: 80
-    targetPort: 80
-    protocol: TCP
-  selector:
-    app: cega-users
