+++ deployments/kube/yml/cega-mq/sts.cega-mq.yml
@@ -2,7 +2,6 @@ apiVersion: apps/v1
 kind: StatefulSet
 metadata:
   name: cega-mq
-  namespace: localega
   labels:
     role: broker
 spec:
