+++ deployments/kube/yml/cega-mq/svc.cega-mq.yml
@@ -2,7 +2,6 @@ apiVersion: v1
 kind: Service
 metadata:
   name: cega-mq-management
-  namespace: localega
   labels:
     app: cega-mq
 spec:
@@ -20,7 +19,6 @@ kind: Service
 metadata:
   # The required headless service for StatefulSets
   name: cega-mq
-  namespace: localega
   labels:
     app: cega-mq
 spec:
