+++ deployments/kube/yml/lega-config/secret.lega.yml
@@ -34,4 +34,4 @@ metadata:
 type: Opaque
 data:
   credentials: <change me>
-  secret_key: <change me>
+  # secret_key: <change me>
