+++ deployments/kube/yml/mq/cm.lega-mq.yml
@@ -5,9 +5,23 @@ metadata:
   namespace: localega
 data:
   rabbitmq.config: |-
-    <rabbitmq.config>
+    %% -*- mode: erlang -*-
+    %%
+    [{rabbit,[{loopback_users, [ ] },
+              {tcp_listeners, [ 5672 ] },
+              {ssl_listeners, [ ] },
+              {hipe_compile, false },
+    	  {default_vhost, "/"},
+    	  {default_user,  "guest"},
+    	  {default_pass,  "guest"},
+    	  {default_permissions, [".*", ".*",".*"]},
+    	  {default_user_tags, [administrator]},
+    	  {disk_free_limit, "1GB"}]},
+     {rabbitmq_management, [ { listener, [ { port, 15672 }, { ssl, false }] },
+                             { load_definitions, "/etc/rabbitmq/defs.json"} ]}
+    ].
   defs.json: |-
-    <defs.json>
+    <change me>
 ---
 apiVersion: v1
 kind: ConfigMap
