+++ deployments/kube/yml/mq/sts.lega-mq.yml
