+++ deployments/kube/yml/mq/svc.lega-mq.yml
