+++ deployments/kube/yml/s3/sc.minio.yml
