+++ deployments/kube/yml/s3/sts.minio.yml
