+++ deployments/kube/yml/s3/svc.minio.yml
