+++ deployments/kube/yml/verify/sts.verify.yml
